--- a/compiler/plugins/target/AMD-AIE/iree-amd-aie/Target/mm_npu1.cc
+++ b/compiler/plugins/target/AMD-AIE/iree-amd-aie/Target/mm_npu1.cc
@@ -472,11 +472,14 @@ extern "C" {
   }
 
 matmul_combos(matmul_vectorized_c_func, 16, 16, 32)
+matmul_combos(matmul_vectorized_c_func, 16, 16, 64)
 matmul_combos(matmul_vectorized_c_func, 32, 32, 32)
 matmul_combos(matmul_vectorized_c_func, 64, 64, 64)
 matmul_combos(matmul_vectorized_c_func, 32, 32, 64)
 matmul_combos(matmul_vectorized_c_func, 32, 32, 128)
+matmul_combos(matmul_vectorized_c_func, 64, 32, 128)
 matmul_combos_i8(matmul_vectorized_c_func, 16, 16, 32)
+matmul_combos_i8(matmul_vectorized_c_func, 16, 16, 64)
 matmul_combos_i8(matmul_vectorized_c_func, 32, 32, 8)
 matmul_combos_i8(matmul_vectorized_c_func, 32, 32, 16)
 matmul_combos_i8(matmul_vectorized_c_func, 32, 32, 32)
@@ -487,6 +490,7 @@ matmul_combos_i8(matmul_vectorized_c_func, 64, 64, 128)
 
 zero_fill_combos(zero_vectorized_c_func, 16, 16)
 zero_fill_combos(zero_vectorized_c_func, 32, 32)
+zero_fill_combos(zero_vectorized_c_func, 64, 32)
 zero_fill_combos(zero_vectorized_c_func, 64, 64)
 
 }  // extern "C"